Deleted user tables, left only getting claims
Since all work with user data lies with the Identity Server, we do not need user tables here
