+++ API/Data/Context.cs
@@ -1,6 +1,4 @@
-﻿using API.Data.Models;
-
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 
 namespace API.Data;
 
@@ -9,7 +7,4 @@ public sealed class Context : DbContext
     public Context(DbContextOptions<Context> options) : base(options)
     {
     }
-
-    public DbSet<User> Users => Set<User>();
-    public DbSet<UserCredential> UserCredentials => Set<UserCredential>();
 }
