+++ API/Program.cs
@@ -54,13 +54,6 @@ services.AddControllers();
 
 var app = builder.Build();
 
-using (var scope = app.Services.CreateScope())
-{
-    var serviceProvider = scope.ServiceProvider;
-
-    SeedData.Initialize(serviceProvider);
-}
-
 app.UseHttpsRedirection();
 
 app.UseCors();
